requirements.txt
-r requirements_local.txt
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
whitenoise==1.0.3
boto==2.32.1
django-storages==1.1.8
